Refactor tests to be integration tests
