+++ checkpointer_test.go
@@ -55,8 +55,8 @@ func TestDoesTableExist(t *testing.T) {
 func TestGetLeaseNotAquired(t *testing.T) {
 	svc := &mockDynamoDB{tableExist: true}
 	checkpoint := &DynamoCheckpoint{
-		TableName: "TableName",
-		svc:       svc,
+		TableName:      "TableName",
+		skipTableCheck: true,
 	}
 	checkpoint.Init()
 	checkpoint.svc = svc
@@ -82,8 +82,10 @@ func TestGetLeaseNotAquired(t *testing.T) {
 func TestGetLeaseAquired(t *testing.T) {
 	svc := &mockDynamoDB{tableExist: true}
 	checkpoint := &DynamoCheckpoint{
-		TableName: "TableName",
+		TableName:      "TableName",
+		skipTableCheck: true,
 	}
+	checkpoint.svc = svc
 	checkpoint.Init()
 	checkpoint.svc = svc
 	marshalledCheckpoint := map[string]*dynamodb.AttributeValue{
@@ -124,81 +126,11 @@ func TestGetLeaseAquired(t *testing.T) {
 	}
 }
 
-func TestRaceCondGetLeaseTimeout(t *testing.T) {
-	svc := &mockDynamoDB{tableExist: true}
-	checkpoint := &DynamoCheckpoint{
-		TableName: "TableName",
-	}
-	checkpoint.Init()
-	checkpoint.svc = svc
-	marshalledCheckpoint := map[string]*dynamodb.AttributeValue{
-		"ShardID": {
-			S: aws.String("0001"),
-		},
-		"AssignedTo": {
-			S: aws.String("abcd-efgh"),
-		},
-		"LeaseTimeout": {
-			S: aws.String(time.Now().AddDate(0, -1, 0).UTC().Format(time.RFC3339)),
-		},
-		"SequenceID": {
-			S: aws.String("deadbeef"),
-		},
-	}
-	input := &dynamodb.PutItemInput{
-		TableName: aws.String("TableName"),
-		Item:      marshalledCheckpoint,
-	}
-	checkpoint.svc.PutItem(input)
-	shard := &shardStatus{
-		ID:         "0001",
-		Checkpoint: "",
-		mux:        &sync.Mutex{},
-	}
-	err := checkpoint.GetLease(shard, "ijkl-mnop")
-
-	if err == nil || err.Error() != ErrLeaseNotAquired {
-		t.Error("Got a lease when checkpoints didn't match. Potentially we stomped on the checkpoint")
-	}
-}
-
-func TestRaceCondGetLeaseNoAssignee(t *testing.T) {
-	svc := &mockDynamoDB{tableExist: true}
-	checkpoint := &DynamoCheckpoint{
-		TableName: "TableName",
-	}
-	checkpoint.Init()
-	checkpoint.svc = svc
-	marshalledCheckpoint := map[string]*dynamodb.AttributeValue{
-		"ShardID": {
-			S: aws.String("0001"),
-		},
-		"SequenceID": {
-			S: aws.String("deadbeef"),
-		},
-	}
-	input := &dynamodb.PutItemInput{
-		TableName: aws.String("TableName"),
-		Item:      marshalledCheckpoint,
-	}
-	checkpoint.svc.PutItem(input)
-	shard := &shardStatus{
-		ID:         "0001",
-		Checkpoint: "",
-		mux:        &sync.Mutex{},
-	}
-	err := checkpoint.GetLease(shard, "ijkl-mnop")
-
-	if err == nil || err.Error() != ErrLeaseNotAquired {
-		t.Error("Got a lease when checkpoints didn't match. Potentially we stomped on the checkpoint")
-	}
-}
-
 func TestGetLeaseRenewed(t *testing.T) {
 	svc := &mockDynamoDB{tableExist: true}
 	checkpoint := &DynamoCheckpoint{
-		TableName: "TableName",
-		svc:       svc,
+		TableName:      "TableName",
+		skipTableCheck: true,
 	}
 	checkpoint.Init()
 	checkpoint.svc = svc
